refac: 함수화
